requirements.txt
SQLAlchemy
aiofiles
aiopg
aiohttp_jinja2
pytest
pytest-aiohttp
pytest-asyncio
pytest-pythonpath